move ApplicationBuilder to "Hosting" namespace
