+++ lib/Hosting/ApplicationBuilder.php
@@ -6,7 +6,7 @@
  * @link        https://github.com/DevNet-Framework
  */
 
-namespace DevNet\Web\Middleware;
+namespace DevNet\Web\Hosting;
 
 use DevNet\System\Async\Task;
 use DevNet\Common\Dependency\IServiceProvider;
@@ -14,6 +14,9 @@ use DevNet\System\MethodTrait;
 use DevNet\System\PropertyTrait;
 use DevNet\Web\Hosting\WebHostEnvironment;
 use DevNet\Web\Http\HttpContext;
+use DevNet\Web\Middleware\IMiddleware;
+use DevNet\Web\Middleware\MiddlewareDelegate;
+use DevNet\Web\Middleware\RequestDelegate;
 
 class ApplicationBuilder implements IApplicationBuilder
 {
