+++ lib/Hosting/WebHostBuilder.php
@@ -15,7 +15,6 @@ use DevNet\Common\Dependency\ServiceProvider;
 use DevNet\System\Runtime\LauncherProperties;
 use DevNet\System\PropertyTrait;
 use DevNet\Web\Http\HttpContext;
-use DevNet\Web\Middleware\ApplicationBuilder;
 use Closure;
 
 class WebHostBuilder implements IWebHostBuilder
